Created training class modules.
