+++ chatterbot/training/trainers.py
@@ -1,15 +1,22 @@
-from .conversation import Statement
-from .corpus import Corpus
+from chatterbot.conversation import Statement
+from chatterbot.corpus import Corpus
 
 
 class Trainer(object):
 
     def __init__(self, storage, **kwargs):
+        self.kwargs = kwargs
         self.storage = storage
         self.corpus = Corpus()
 
-    def train_from_list(self, conversation):
-        recent_statements = []
+    def train(self):
+        pass
+
+
+class ListTrainer(Trainer):
+
+    def train(self, conversation):
+        statement_history = []
 
         for text in conversation:
             statement = self.storage.find(text)
@@ -19,18 +26,23 @@ class Trainer(object):
                 statement = Statement(text)
 
             previous_statement = None
-            if recent_statements:
-                previous_statement = recent_statements[-1]
+            if statement_history:
+                previous_statement = statement_history[-1]
 
             if previous_statement:
                 statement.add_response(previous_statement)
 
-            recent_statements.append(statement)
+            statement_history.append(statement)
             self.storage.update(statement)
 
-    def train_from_corpora(self, corpora):
+
+class ChatterBotCorpusTrainer(Trainer):
+
+    def train(self, *corpora):
+        trainer = ListTrainer(self.storage)
+
         for corpus in corpora:
             corpus_data = self.corpus.load_corpus(corpus)
             for data in corpus_data:
                 for pair in data:
-                    self.train_from_list(pair)
+                    trainer.train(pair)
